+++ changelog.md
@@ -1,5 +1,10 @@
 # Version history of weather-information-collector
 
+## Version 0.9.7-pre (2019-10-??)
+
+The program is now able to collect data from the Weatherstack API, too.
+However, only data about the current weather can be collected (yet).
+
 ## Version 0.9.6 (2019-09-22)
 
 Synchronization program will not start the synchronization, if both the source